Update 1.2.3
Working scene just use te stuff and scene loading is left

+++ RawEngine/core/Camera.h
@@ -23,9 +23,9 @@ namespace core {
 		bool enableRotation = true;
 
 		//deltatimers;
-		float finishFrameTime;
-		float	deltaTime;
-		float	currentTime;
+		double finishFrameTime;
+		double	deltaTime;
+		double	currentTime;
